Add dielectric sphere with impedance layer
The dielectric sphere has an impedance layer. As a consequence,
the potential is discontinuous at the interface, but the radial
displacement field is still continuous.

This approximation is commonly used in BioEM to model cells with
their resistive membranes.

The approximation is useful to observe convergence of numerical
algorithms that use a similar thin layer approximation (and
which would not converge to the full three layer problem)